+++ aiotfm/__init__.py
@@ -1,10 +1,9 @@
-from aiotfm import utils, errors
+from aiotfm import utils, errors, enums
 from aiotfm.client import Client
 from aiotfm.connection import Connection
 from aiotfm.packet import Packet
 from aiotfm.player import Player, Profile, Stats
 from aiotfm.tribe import *
-from aiotfm.utils import enums
 
 from aiotfm.__version__ import __author__, __title__, __description__
 from aiotfm.__version__ import __url__, __version__, __credits__, __license__
